Add support for built-in per-task rerun option
Issue  #9166 description provides a great summary, but the primary
reason is to allow users to rerun specific tasks (that would otherwise
not be run), like test, without incurring the overhead of re-running all
other tasks (be it for the sake of time saving, or such as when
benchmarking a specific task, such as in #20641).

Additionally, when Gradle has a cache corruption issue (see “Corrupt
plugin jars end up in Gradle's local caches #20830”), rerunning the
affected tasks is the easiest and finest-grained workaround instead of
messing up with local cache state (or resorting to a full re-run of the
task graph).

Issues: #9166, #20641